<!--

Thank you for sending a pull request! Here are some tips:

1. If this is your first time, please read our contribution guide at https://github.com/grafana/grafana/blob/main/CONTRIBUTING.md

2. Ensure you include and run the appropriate tests as part of your Pull Request.

3. In a new feature or configuration option, an update to the documentation is necessary. Everything related to the documentation is under the docs folder in the root of the repository.

4. If the Pull Request is a work in progress, make use of GitHub's "Draft PR" feature and mark it as such.

5. If you can not merge your Pull Request due to a merge conflict, Rebase it. This gets it in sync with the main branch.

6. Name your PR as "<FeatureArea>: Describe your change", e.g. Alerting: Prevent race condition. If it's a fix or feature relevant for the changelog describe the user impact in the title. The PR title is used to auto-generate the changelog for issues marked with the "add to changelog" label.

-->

**What this PR does / why we need it**:

old == `bluge.InMemoryOnlyConfig()` vs new == `bluge.DefaultConfig("bluge.db"))`

```
name                                        old time/op    new time/op    delta
_DashboardIndexUpdates/dashboard-delete-32    20.2µs ±10%  6577.3µs ± 1%  +32418.54%  (p=0.000 n=10+8)
_DashboardIndexUpdates/dashboard-update-32    2.24ms ± 3%   19.91ms ± 2%    +789.01%  (p=0.000 n=9+10)
_DashboardIndexUpdates/reindex-32             11.3ms ± 7%    61.9ms ± 4%    +447.01%  (p=0.000 n=10+9)

name                                        old alloc/op   new alloc/op   delta
_DashboardIndexUpdates/dashboard-delete-32    6.68kB ± 0%   70.43kB ±34%    +953.80%  (p=0.000 n=10+10)
_DashboardIndexUpdates/dashboard-update-32    1.59MB ± 3%    0.44MB ±17%     -72.53%  (p=0.000 n=10+10)
_DashboardIndexUpdates/reindex-32             7.49MB ± 1%    8.49MB ±20%     +13.27%  (p=0.003 n=10+10)

name                                        old allocs/op  new allocs/op  delta
_DashboardIndexUpdates/dashboard-delete-32       110 ± 0%       825 ±23%    +650.00%  (p=0.000 n=10+8)
_DashboardIndexUpdates/dashboard-update-32     9.03k ± 3%     3.79k ±31%     -58.05%  (p=0.000 n=10+10)
_DashboardIndexUpdates/reindex-32              82.1k ± 0%    114.5k ±43%     +39.50%  (p=0.000 n=9+10)
```

**Which issue(s) this PR fixes**:

<!--

- Automatically closes linked issue when the Pull Request is merged.

Usage: "Fixes #<issue number>", or "Fixes (paste link of issue)"

-->

Fixes #

**Special notes for your reviewer**: